requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
matplotlib>=3.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
pillow>=9.0.0
tqdm>=4.60.0
seaborn>=0.11.0
tensorboard>=2.8.0
opencv-python>=4.6.0
albumentations>=1.1.0 